:hammer: Restructure file hierarchy
