+++ src/baseSimulator.h
@@ -3,26 +3,36 @@
  */
 #pragma once
 
-#include "simulation/CFDSim.h"
-#include "simulation/Droplet.h"
-#include "simulation/Fluid.h"
-#include "simulation/Injection.h"
-#include "simulation/MembraneModels.h"
-#include "simulation/Mixture.h"
-#include "simulation/MixtureInjection.h"
-#include "simulation/MixingModels.h"
-#include "simulation/ResistanceModels.h"
-#include "simulation/Simulation.h"
-#include "simulation/simulators/cfdSimulator.h"
-#include "simulation/simulators/olbContinuous.h"
-#include "simulation/simulators/olbMixing.h"
-#include "simulation/simulators/olbOoc.h"
-#include "simulation/Specie.h"
-#include "simulation/Tissue.h"
+#include "simulation/entities/Droplet.h"
+#include "simulation/entities/Fluid.h"
+#include "simulation/entities/Mixture.h"
+#include "simulation/entities/Specie.h"
+#include "simulation/entities/Tissue.h"
+
 #include "simulation/events/BoundaryEvent.h"
 #include "simulation/events/Event.h"
 #include "simulation/events/InjectionEvent.h"
 #include "simulation/events/MergingEvent.h"
+
+#include "simulation/operations/Injection.h"
+#include "simulation/operations/MixtureInjection.h"
+
+#include "simulation/models/MembraneModels.h"
+#include "simulation/models/MixingModels.h"
+#include "simulation/models/ResistanceModels.h"
+
+#include "simulation/simulators/Simulation.h"
+#include "simulation/simulators/AbstractContinuous.h"
+#include "simulation/simulators/AbstractDroplet.h"
+#include "simulation/simulators/AbstractMixing.h"
+#include "simulation/simulators/AbstractMembrane.h"
+#include "simulation/simulators/HybridContinuous.h"
+
+#include "simulation/simulators/CFDSim.h"
+#include "simulation/simulators/cfdHandlers/cfdSimulator.h"
+#include "simulation/simulators/cfdHandlers/olbContinuous.h"
+// #include "simulation/simulators/cfdHandlers/olbMixing.h" //** TODO: HybridMixingSimulation */
+// #include "simulation/simulators/cfdHandlers/olbOoc.h"    //** TODO: HybridOocSimulation */
 
 #include "nodalAnalysis/NodalAnalysis.h"
