+++ src/simulation/entities/Droplet.h
