+++ src/simulation/models/CMakeLists.txt
@@ -1,0 +1,15 @@
+set(SOURCE_LIST
+    MembraneModels.hh
+    MixingModels.hh
+    ResistanceModels.hh
+)
+
+set(HEADER_LIST
+    MembraneModels.h
+    MixingModels.h
+    ResistanceModels.h
+)
+
+target_sources(${TARGET_NAME} PUBLIC ${SOURCE_LIST} ${HEADER_LIST})
+target_include_directories(${TARGET_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_link_libraries(${TARGET_NAME} PUBLIC lbmLib)
