+++ src/simulation/models/ResistanceModels.h
