+++ src/simulation/simulators/CMakeLists.txt
@@ -1,21 +1,25 @@
 set(SOURCE_LIST
-    cfdSimulator.hh
-    olbContinuous.hh
-    olbMixing.hh
-    olbOoc.hh
+    Simulation.hh
+    AbstractContinuous.hh
+    AbstractDroplet.hh
+    AbstractMixing.hh
+    AbstractMembrane.hh
+    CFDSim.hh
+    HybridContinuous.hh
 )
 
 set(HEADER_LIST
-    cfdSimulator.h
-    olbContinuous.h
-    olbMixing.h
-    olbOoc.h
+    Simulation.h
+    AbstractContinuous.h
+    AbstractDroplet.h
+    AbstractMixing.h
+    AbstractMembrane.h
+    CFDSim.h
+    HybridContinuous.h
 )
 
-if(USE_ESSLBM)
-    target_sources(${TARGET_NAME} PUBLIC essContinuous.hh essContinuous.h)
-    target_sources(${TARGET_NAME} PUBLIC essMixing.hh essMixing.h)
-endif()
+target_sources(${TARGET_NAME} PUBLIC ${SOURCE_LIST} ${HEADER_LIST})
+target_include_directories(${TARGET_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_link_libraries(${TARGET_NAME} PUBLIC lbmLib)
 
-target_sources(${TARGET_NAME} PRIVATE ${SOURCE_LIST} ${HEADER_LIST})
-target_include_directories(${TARGET_NAME} PRIVATE ${CMAKE_CURRENT_LIST_DIR})
+add_subdirectory(cfdHandlers)
